requirements.txt
affine==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
cftime==1.6.4.post1
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
coverage==7.10.0
dash==3.2.0
dash-bootstrap-components==2.0.3
Flask==3.1.1
frozenlist==1.7.0
geopandas==1.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
narwhals==1.48.1
nest-asyncio==1.6.0
netCDF4==1.7.2
numpy==2.3.2
packaging==25.0
pandas==2.3.1
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
pooch==1.8.2
pprintpp==0.4.0
propcache==0.3.2
pycountry==24.6.1
Pygments==2.19.2
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
repoze.lru==0.7
requests==2.32.4
retrying==1.4.1
scipy==1.16.1
setuptools==78.1.1
shapely==2.1.1
six==1.17.0
SRTM.py==0.3.7
tiny-retriever==0.1.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
xarray==2025.7.1
yarl==1.20.1
zipp==3.23.0
